fast forward to 1.3.0 from fabric

+++ src/main/kotlin/us/timinc/mc/cobblemon/unchained/api/AbstractActionInfluenceBooster.kt
@@ -35,7 +35,7 @@ abstract class AbstractActionInfluenceBooster(
         pokemon: Pokemon,
         species: ResourceLocation,
         form: String,
-        points: Double,
+        points: Float,
         player: ServerPlayer,
     )
 }
